refactor(cli): split orchestrate.py into package (fanout/flow/_common)
orchestrate.py (1610 lines) → orchestrate/ package:
- _common.py (163): shared models, formatting, team helpers, worker resolution
- fanout.py (485): fan-out execution
- flow.py (719): DAG flow with control nodes
- __init__.py (287): CLI registration and dispatch

No behavioral changes — pure file reorganization.
