--- lionagi/cli/orchestrate.py
@@ -1,1610 +1,0 @@
-# Copyright (c) 2023-2026, HaiyangLi <quantocean.li at gmail dot com>
-# SPDX-License-Identifier: Apache-2.0
-"""`li orchestrate fanout` — parallel multi-agent fan-out with optional synthesis.
-
-Three-phase execution:
-  1. Orchestrator decomposes task → generates N AgentRequests (instruct + model)
-  2. Fan out → each request runs as a communicate node
-  3. Synthesis → orchestrator combines all results
-
-Examples:
-    # Pattern A: homogeneous, no synthesis
-    li o fanout codex/gpt-5.4-xhigh "audit for dead code" -n 3
-
-    # Pattern B: heterogeneous workers, strong model synthesizes
-    li o fanout claude/opus-4-6-medium "analyze auth middleware" \\
-        --workers "codex/gpt-5.4-high, claude/sonnet, codex/gpt-5.3-codex" \\
-        --with-synthesis
-
-    # Pattern C: homogeneous workers, different synthesis model
-    li o fanout claude/sonnet "suggest approaches" \\
-        -n 3 --with-synthesis claude/opus-4-6-medium
-"""
-
-from __future__ import annotations
-
-import argparse
-import sys
-import time
-from pathlib import Path
-
-from lionagi import Branch, FieldModel, Session, iModel, json_dumps
-from lionagi._errors import TimeoutError as LionTimeoutError
-from lionagi.ln import acreate_path
-from lionagi.ln.concurrency import move_on_after, run_async
-from lionagi.models import HashableModel
-from lionagi.operations.builder import OperationGraphBuilder
-from lionagi.operations.fields import Instruct
-from lionagi.protocols.generic.log import DataLoggerConfig
-
-from ._agents import AgentProfile, load_agent_profile
-from ._persistence import LIONAGI_HOME, save_last_branch_pointer
-from ._providers import add_common_cli_args, build_imodel_from_spec, parse_model_spec
-from .team import _load_team, _now_iso, _save_team, _teams_dir
-
-# ── Agent request model (structured output from orchestrator) ─────────────
-
-
-class AgentRequest(HashableModel):
-    """A single agent request generated by the orchestrator."""
-
-    instruct: Instruct
-    model: str | None = None
-
-
-AGENT_REQUEST_FIELDS = FieldModel(list[AgentRequest], name="agents")
-
-
-# ── Output formatting ─────────────────────────────────────────────────────
-
-
-def _format_result_text(
-    worker_results: list[dict],
-    synthesis_result: dict | None = None,
-) -> str:
-    lines = []
-    for i, w in enumerate(worker_results, 1):
-        n = len(worker_results)
-        lines.append(f"{'═' * 60}")
-        lines.append(f"  Worker {i}/{n}  [{w['model']}]")
-        lines.append(f"  {w['time_ms']:.0f}ms")
-        lines.append(f"{'═' * 60}")
-        lines.append(w.get("response", "(no response)"))
-        lines.append("")
-
-    if synthesis_result is not None:
-        lines.append(f"{'═' * 60}")
-        lines.append(f"  Synthesis  [{synthesis_result['model']}]")
-        lines.append(f"  {synthesis_result['time_ms']:.0f}ms")
-        lines.append(f"{'═' * 60}")
-        lines.append(synthesis_result.get("response", "(no response)"))
-        lines.append("")
-
-    return "\n".join(lines)
-
-
-def _format_result_json(
-    worker_results: list[dict],
-    synthesis_result: dict | None = None,
-) -> str:
-    out = {"workers": worker_results}
-    if synthesis_result is not None:
-        out["synthesis"] = synthesis_result
-    return json_dumps(out)
-
-
-# ── Team-mode guidance ────────────────────────────────────────────────────
-
-TEAM_WORKER_SYSTEM = """\
-You are **{worker_name}**, a specialist on team "{team_name}" (id: {team_id}).
-
-## Your team
-{roster_text}
-
-## How this works
-1. You receive a focused assignment from the orchestrator.
-2. Work independently — produce clear, actionable output.
-3. Your results are automatically shared with the team via `li team`.
-4. After this round, teammates or the orchestrator can follow up:
-   - `li team receive -t {team_id} --as {worker_name}` to read messages
-   - `li team send "..." -t {team_id} --to {worker_name}` to reply
-   - `li agent -r {{branch_id}} "follow-up"` to continue your session
-
-Focus on YOUR assignment. Be specific — your output is the deliverable.\
-"""
-
-
-def _create_fanout_team(
-    team_name: str,
-    worker_names: list[str],
-) -> dict:
-    """Create a team for fanout, returning the team data dict."""
-    from uuid import uuid4
-
-    team_id = uuid4().hex[:12]
-    members = ["orchestrator"] + worker_names
-    data = {
-        "id": team_id,
-        "name": team_name,
-        "members": members,
-        "messages": [],
-        "created_at": _now_iso(),
-    }
-    _save_team(data)
-    return data
-
-
-def _post_results_to_team(
-    team_data: dict,
-    worker_results: list[dict],
-    worker_names: list[str],
-    synthesis_result: dict | None = None,
-) -> None:
-    """Post worker outputs and synthesis as team messages."""
-    from uuid import uuid4
-
-    for wr, name in zip(worker_results, worker_names):
-        team_data["messages"].append(
-            {
-                "id": uuid4().hex[:12],
-                "from": name,
-                "to": ["*"],
-                "content": wr.get("response", "(no response)"),
-                "timestamp": _now_iso(),
-                "read_by": [],
-            }
-        )
-
-    if synthesis_result:
-        team_data["messages"].append(
-            {
-                "id": uuid4().hex[:12],
-                "from": "orchestrator",
-                "to": ["*"],
-                "content": f"[SYNTHESIS]\n{synthesis_result.get('response', '')}",
-                "timestamp": _now_iso(),
-                "read_by": [],
-            }
-        )
-
-    _save_team(team_data)
-
-
-# ── Core fanout execution ─────────────────────────────────────────────────
-
-
-def _resolve_worker_spec(
-    token: str,
-) -> tuple[str, AgentProfile | None]:
-    """Resolve a worker token to (model_spec, profile_or_None).
-
-    If token contains '/' → model spec (e.g. 'codex/gpt-5.4').
-    Otherwise → try loading as agent profile name.
-    """
-    if "/" in token:
-        return token, None
-    try:
-        profile = load_agent_profile(token)
-        return profile.model or "codex/gpt-5.4", profile
-    except FileNotFoundError:
-        return token, None
-
-
-async def _run_fanout(
-    model_spec: str,
-    prompt: str,
-    *,
-    num_workers: int = 3,
-    workers_str: str | None = None,
-    with_synthesis: bool = False,
-    synthesis_model: str | None = None,
-    synthesis_prompt: str | None = None,
-    max_concurrent: int = 0,
-    yolo: bool = False,
-    verbose: bool = False,
-    effort: str | None = None,
-    theme: str | None = None,
-    output_format: str = "text",
-    save_dir: str | None = None,
-    team_name: str | None = None,
-    cwd: str | None = None,
-    timeout: int | None = None,
-    agent_name: str | None = None,
-) -> str:
-    """Three-phase fan-out: decompose → fan out → synthesize."""
-    if timeout:
-        with move_on_after(timeout) as cancel_scope:
-            result = await _run_fanout_inner(
-                model_spec,
-                prompt,
-                num_workers=num_workers,
-                workers_str=workers_str,
-                with_synthesis=with_synthesis,
-                synthesis_model=synthesis_model,
-                synthesis_prompt=synthesis_prompt,
-                max_concurrent=max_concurrent,
-                yolo=yolo,
-                verbose=verbose,
-                effort=effort,
-                theme=theme,
-                output_format=output_format,
-                save_dir=save_dir,
-                team_name=team_name,
-                cwd=cwd,
-                agent_name=agent_name,
-            )
-        if cancel_scope.cancelled_caught:
-            raise LionTimeoutError(f"Fanout timed out after {timeout}s")
-        return result
-    return await _run_fanout_inner(
-        model_spec,
-        prompt,
-        num_workers=num_workers,
-        workers_str=workers_str,
-        with_synthesis=with_synthesis,
-        synthesis_model=synthesis_model,
-        synthesis_prompt=synthesis_prompt,
-        max_concurrent=max_concurrent,
-        yolo=yolo,
-        verbose=verbose,
-        effort=effort,
-        theme=theme,
-        output_format=output_format,
-        save_dir=save_dir,
-        team_name=team_name,
-        cwd=cwd,
-        agent_name=agent_name,
-    )
-
-
-async def _run_fanout_inner(
-    model_spec: str,
-    prompt: str,
-    *,
-    num_workers: int = 3,
-    workers_str: str | None = None,
-    with_synthesis: bool = False,
-    synthesis_model: str | None = None,
-    synthesis_prompt: str | None = None,
-    max_concurrent: int = 0,
-    yolo: bool = False,
-    verbose: bool = False,
-    effort: str | None = None,
-    theme: str | None = None,
-    output_format: str = "text",
-    save_dir: str | None = None,
-    team_name: str | None = None,
-    cwd: str | None = None,
-    agent_name: str | None = None,
-) -> str:
-    """Inner fanout logic (no timeout wrapper)."""
-    t0 = time.monotonic()
-
-    # Load orchestrator agent profile if specified
-    orc_profile: AgentProfile | None = None
-    if agent_name:
-        orc_profile = load_agent_profile(agent_name)
-        if orc_profile.model and not model_spec:
-            model_spec = orc_profile.model
-        if orc_profile.effort and not effort:
-            effort = orc_profile.effort
-        if orc_profile.yolo and not yolo:
-            yolo = True
-
-    if not model_spec:
-        raise ValueError(
-            "Provide a model spec or use -a/--agent to load a profile with a model."
-        )
-
-    # Build orchestrator model
-    orc_imodel = build_imodel_from_spec(
-        model_spec,
-        yolo=yolo,
-        verbose=verbose,
-        effort_override=effort,
-        theme=theme,
-    )
-
-    # Determine available workers — resolve agent profile names
-    worker_profiles: list[AgentProfile | None] = []
-    if workers_str:
-        worker_model_list = []
-        for token in (s.strip() for s in workers_str.split(",")):
-            wmodel, wprofile = _resolve_worker_spec(token)
-            worker_model_list.append(wmodel)
-            worker_profiles.append(wprofile)
-    else:
-        ms = parse_model_spec(model_spec)
-        worker_model_list = [ms.model] * num_workers
-        worker_profiles = [None] * num_workers
-
-    # ── Team setup (if --team-mode) ─────────────────────────────────
-    team_data = None
-    worker_names = []
-    for i, wp in enumerate(worker_profiles):
-        if wp and wp.name:
-            base = wp.name
-            count = sum(
-                1 for n in worker_names if n == base or n.startswith(f"{base}-")
-            )
-            worker_names.append(f"{base}-{count + 1}" if count > 0 else base)
-        else:
-            worker_names.append(f"worker-{i + 1}")
-    if team_name:
-        team_data = _create_fanout_team(team_name, worker_names)
-        if not verbose:
-            print(
-                f"Team '{team_name}' created ({team_data['id']}): "
-                f"{', '.join(worker_names)}",
-                file=sys.stderr,
-            )
-
-    # Apply cwd to endpoint if specified
-    if cwd:
-        orc_imodel.endpoint.config.kwargs.setdefault("repo", Path(cwd))
-
-    # ── Phase 1: Orchestrator decomposes task ─────────────────────────
-    builder = OperationGraphBuilder("Fanout")
-    orc_system = orc_profile.system_prompt if orc_profile else None
-    orc_branch = Branch(
-        chat_model=orc_imodel,
-        system=orc_system,
-        log_config=DataLoggerConfig(auto_save_on_exit=False),
-        name="orchestrator",
-    )
-    session = Session(default_branch=orc_branch)
-
-    # Build guidance with role names when workers are agent profiles
-    worker_descriptions = []
-    for i, wm in enumerate(worker_model_list):
-        wp = worker_profiles[i] if i < len(worker_profiles) else None
-        if wp and wp.name:
-            worker_descriptions.append(
-                f"{worker_names[i]} (role: {wp.name}, model: {wm})"
-            )
-        else:
-            worker_descriptions.append(f"{worker_names[i]} (model: {wm})")
-    roster_guidance = "; ".join(worker_descriptions)
-
-    root = builder.add_operation(
-        "operate",
-        branch=orc_branch,
-        instruct=Instruct(
-            instruction=(
-                f"Generate {len(worker_model_list)} agent requests to address "
-                f"the following task. Each agent should tackle a distinct angle "
-                f"or perspective. Each agent is a WORKER that will DIRECTLY "
-                f"answer its assigned sub-task — NOT delegate further."
-            ),
-            context={"user_prompt": prompt},
-            guidance=(
-                f"Available workers: {roster_guidance}. "
-                f"CRITICAL: You MUST produce your output ONLY via the structured "
-                f"output fields (the AgentRequest list). Do NOT use any provider-native "
-                f"subagent or tool-spawning features (no Agent tool, no subprocess "
-                f"spawning, no delegation tools). The ONLY correct way to dispatch "
-                f"workers is by filling in the structured AgentRequest output. "
-                f"Each AgentRequest instruction must be a DIRECT task the worker "
-                f"will answer itself — not a meta-instruction to generate more "
-                f"agents or decompose further. Workers are leaf executors, not "
-                f"orchestrators. "
-                f"Match each sub-task to the worker's role strengths. "
-                f"If multiple models are provided, set the model field to "
-                f"the exact model string for that worker. "
-                f"If all workers use the same model, model can be null."
-            ),
-        ),
-        field_models=[AGENT_REQUEST_FIELDS],
-        reason=True,
-    )
-
-    if not verbose:
-        print(
-            f"Phase 1: Orchestrator decomposing task into "
-            f"{len(worker_model_list)} agent requests...",
-            file=sys.stderr,
-        )
-
-    result1 = await session.flow(builder.get_graph())
-    t_decompose = time.monotonic() - t0
-
-    # Extract agent requests
-    root_result = result1.get("operation_results", {}).get(root)
-    agents: list[AgentRequest] = getattr(root_result, "agents", None) or []
-
-    if not agents:
-        return "Orchestrator produced no agent requests."
-
-    if not verbose:
-        print(
-            f"Phase 1 done ({t_decompose:.1f}s): {len(agents)} requests generated.",
-            file=sys.stderr,
-        )
-
-    # ── Phase 2: Fan out ──────────────────────────────────────────────
-    default_ms = parse_model_spec(model_spec)
-    fanned_nodes = []
-    fanned_labels = []
-
-    for i, a in enumerate(agents):
-        wprofile = worker_profiles[i] if i < len(worker_profiles) else None
-        worker_model = a.model or default_ms.model
-        if wprofile and wprofile.model:
-            worker_model = wprofile.model
-        w_effort = effort
-        if wprofile and wprofile.effort and not effort:
-            w_effort = wprofile.effort
-        w_yolo = yolo
-        if wprofile and wprofile.yolo:
-            w_yolo = True
-        worker_imodel = build_imodel_from_spec(
-            worker_model,
-            yolo=w_yolo,
-            verbose=verbose,
-            effort_override=w_effort,
-            theme=theme,
-        )
-        if cwd:
-            worker_imodel.endpoint.config.kwargs.setdefault("repo", Path(cwd))
-        wname = worker_names[i]
-        if team_data:
-            teammates = [n for n in worker_names if n != wname]
-            roster_lines = [f"- orchestrator (coordinator)"]
-            roster_lines += [f"- {t}" for t in teammates]
-            roster_lines.append(f"- **{wname}** (you)")
-            worker_system = TEAM_WORKER_SYSTEM.format(
-                worker_name=wname,
-                team_name=team_data["name"],
-                team_id=team_data["id"],
-                roster_text="\n".join(roster_lines),
-            )
-        elif wprofile and wprofile.system_prompt:
-            worker_system = wprofile.system_prompt
-        else:
-            worker_system = (
-                "You are a leaf worker agent. Complete your assigned task "
-                "directly. You may read files, use tools, and run commands "
-                "as needed. Do NOT spawn sub-agents or delegate further."
-            )
-        worker_branch = Branch(
-            chat_model=worker_imodel,
-            system=worker_system,
-            log_config=DataLoggerConfig(auto_save_on_exit=False),
-            name=wname,
-        )
-        session.include_branches(worker_branch)
-        worker_context = [
-            {"overall_task": prompt},
-            a.instruct.context or "",
-        ]
-        node = builder.add_operation(
-            "communicate",
-            branch=worker_branch,
-            depends_on=[root],
-            instruction=a.instruct.instruction,
-            guidance=a.instruct.guidance,
-            context=worker_context,
-        )
-        fanned_nodes.append(node)
-        fanned_labels.append(worker_model)
-
-    if not verbose:
-        labels = ", ".join(fanned_labels)
-        print(
-            f"Phase 2: Fanning out to {len(fanned_nodes)} workers: [{labels}]",
-            file=sys.stderr,
-        )
-
-    t1 = time.monotonic()
-    conc = max_concurrent if max_concurrent > 0 else len(fanned_nodes)
-    result2 = await session.flow(
-        builder.get_graph(),
-        max_concurrent=conc,
-        verbose=verbose,
-    )
-    t_fanout = time.monotonic() - t1
-
-    # Collect results
-    op_results = result2.get("operation_results", {})
-    worker_results = []
-    contexts = []
-    for i, nid in enumerate(fanned_nodes):
-        res = op_results.get(nid)
-        response_text = str(res) if res is not None else "(no response)"
-        worker_results.append(
-            {
-                "worker": i + 1,
-                "model": fanned_labels[i],
-                "response": response_text,
-                "time_ms": t_fanout * 1000,
-            }
-        )
-        contexts.append(response_text)
-
-    if not verbose:
-        print(f"Phase 2 done ({t_fanout:.1f}s).", file=sys.stderr)
-
-    # ── Phase 3: Synthesis ────────────────────────────────────────────
-    synthesis_result = None
-    if with_synthesis and contexts:
-        synth_spec = synthesis_model or model_spec
-        synth_label = str(parse_model_spec(synth_spec))
-
-        if not verbose:
-            print(f"Phase 3: Synthesis [{synth_label}]...", file=sys.stderr)
-
-        synth_instruction = synthesis_prompt or (
-            f"Synthesize the following {len(contexts)} worker responses "
-            f"into a cohesive analysis.\n\n"
-            f"Original task: {prompt}"
-        )
-
-        synth_node = builder.add_operation(
-            "communicate",
-            branch=orc_branch,
-            depends_on=fanned_nodes,
-            instruction=synth_instruction,
-            context=contexts,
-        )
-
-        t2 = time.monotonic()
-        result3 = await session.flow(builder.get_graph(), verbose=verbose)
-        t_synth = time.monotonic() - t2
-
-        synth_res = result3.get("operation_results", {}).get(synth_node)
-        synthesis_result = {
-            "model": synth_label,
-            "response": str(synth_res) if synth_res is not None else "(no response)",
-            "time_ms": t_synth * 1000,
-        }
-
-        if not verbose:
-            print(f"Phase 3 done ({t_synth:.1f}s).", file=sys.stderr)
-
-    # ── Output ────────────────────────────────────────────────────────
-    if output_format == "json":
-        output = _format_result_json(worker_results, synthesis_result)
-    else:
-        output = _format_result_text(worker_results, synthesis_result)
-
-    # ── Save ──────────────────────────────────────────────────────────
-    if save_dir:
-        save_path = Path(save_dir)
-        save_path.mkdir(parents=True, exist_ok=True)
-        for wr in worker_results:
-            p = save_path / f"worker_{wr['worker']}.md"
-            p.write_text(wr["response"])
-        if synthesis_result:
-            (save_path / "synthesis.md").write_text(synthesis_result["response"])
-        meta = {
-            "prompt": prompt,
-            "workers": fanned_labels,
-            "synthesis_model": synthesis_result["model"] if synthesis_result else None,
-            "total_time_ms": (time.monotonic() - t0) * 1000,
-        }
-        (save_path / "meta.json").write_text(json_dumps(meta))
-        if not verbose:
-            print(f"Saved to {save_path}", file=sys.stderr)
-
-    # ── Post to team ─────────────────────────────────────────────────
-    if team_data:
-        _post_results_to_team(team_data, worker_results, worker_names, synthesis_result)
-        if not verbose:
-            print(
-                f"\nTeam '{team_data['name']}' ({team_data['id']}): "
-                f"{len(worker_results)} results posted.",
-                file=sys.stderr,
-            )
-            print(
-                f"  li team receive -t {team_data['id']} --as orchestrator",
-                file=sys.stderr,
-            )
-            print(
-                f"  li team show {team_data['id']}",
-                file=sys.stderr,
-            )
-
-    # ── Persist all branches ─────────────────────────────────────────
-    orc_provider = orc_branch.chat_model.endpoint.config.provider
-    orc_branch_id = str(orc_branch.id)
-
-    # Save orchestrator
-    orc_path = await acreate_path(
-        directory=LIONAGI_HOME / "logs" / "agents" / orc_provider,
-        filename=orc_branch_id,
-        file_exist_ok=True,
-    )
-    await orc_path.write_text(json_dumps(orc_branch.to_dict()))
-    save_last_branch_pointer(orc_provider, orc_branch_id)
-
-    # Save all worker branches (everything in session except orchestrator)
-    worker_branch_ids = []
-    for branch in session.branches:
-        if str(branch.id) == orc_branch_id:
-            continue
-        w_provider = branch.chat_model.endpoint.config.provider
-        w_branch_id = str(branch.id)
-        w_path = await acreate_path(
-            directory=LIONAGI_HOME / "logs" / "agents" / w_provider,
-            filename=w_branch_id,
-            file_exist_ok=True,
-        )
-        await w_path.write_text(json_dumps(branch.to_dict()))
-        worker_branch_ids.append((w_provider, w_branch_id))
-
-    t_total = time.monotonic() - t0
-    if not verbose:
-        print(f"\nTotal: {t_total:.1f}s", file=sys.stderr)
-
-    print(f'\n[orchestrator] li agent -r {orc_branch_id} "..."', file=sys.stderr)
-    for i, (wp, wid) in enumerate(worker_branch_ids):
-        print(f'[worker-{i+1}]      li agent -r {wid} "..."', file=sys.stderr)
-
-    return output
-
-
-# ── Flow: DAG pipeline execution ──────────────────────────────────────────
-
-
-class FlowAgentSpec(HashableModel):
-    """One agent in a DAG flow — dependencies expressed via depends_on."""
-
-    id: str
-    role: str
-    instruction: str
-    depends_on: list[str] | None = None
-    control: bool = False
-    guidance: str | None = None
-    model: str | None = None
-
-
-class FlowPlan(HashableModel):
-    """DAG execution plan — agents reference each other via id/depends_on.
-
-    Agents with control=true are flow control checkpoints. They receive
-    all prior results and produce a structured verdict (should_continue).
-    If should_continue=true, the orchestrator plans the next batch.
-    """
-
-    agents: list[FlowAgentSpec]
-    synthesis: bool = False
-
-
-FLOW_PLAN_FIELDS = FieldModel(FlowPlan, name="plan")
-
-
-class FlowControlVerdict(HashableModel):
-    """Structured output from a flow control node."""
-
-    should_continue: bool
-    reason: str
-    next_steps: str | None = None
-
-
-FLOW_VERDICT_FIELDS = FieldModel(FlowControlVerdict, name="verdict")
-
-
-def _format_flow_result_text(
-    agent_results: list[dict],
-    synthesis_result: dict | None = None,
-) -> str:
-    lines = []
-    for w in agent_results:
-        deps = w.get("depends_on") or []
-        dep_str = f"  deps: {', '.join(deps)}" if deps else ""
-        lines.append(f"{'═' * 60}")
-        lines.append(f"  {w['id']} ({w['name']})  [{w['model']}]{dep_str}")
-        lines.append(f"  {w['time_ms']:.0f}ms")
-        lines.append(f"{'═' * 60}")
-        lines.append(w.get("response", "(no response)"))
-        lines.append("")
-
-    if synthesis_result is not None:
-        lines.append(f"{'═' * 60}")
-        lines.append(f"  Synthesis  [{synthesis_result['model']}]")
-        lines.append(f"  {synthesis_result['time_ms']:.0f}ms")
-        lines.append(f"{'═' * 60}")
-        lines.append(synthesis_result.get("response", "(no response)"))
-        lines.append("")
-
-    return "\n".join(lines)
-
-
-async def _run_flow(
-    model_spec: str,
-    prompt: str,
-    *,
-    with_synthesis: bool = False,
-    synthesis_model: str | None = None,
-    max_concurrent: int = 0,
-    yolo: bool = False,
-    verbose: bool = False,
-    effort: str | None = None,
-    theme: str | None = None,
-    output_format: str = "text",
-    save_dir: str | None = None,
-    team_name: str | None = None,
-    cwd: str | None = None,
-    timeout: int | None = None,
-    agent_name: str | None = None,
-    bare: bool = False,
-    max_agents: int = 0,
-    dry_run: bool = False,
-) -> str:
-    """Auto-DAG flow: orchestrator plans DAG → engine executes with deps."""
-    if timeout:
-        with move_on_after(timeout) as cancel_scope:
-            result = await _run_flow_inner(
-                model_spec,
-                prompt,
-                with_synthesis=with_synthesis,
-                synthesis_model=synthesis_model,
-                max_concurrent=max_concurrent,
-                yolo=yolo,
-                verbose=verbose,
-                effort=effort,
-                theme=theme,
-                output_format=output_format,
-                save_dir=save_dir,
-                team_name=team_name,
-                cwd=cwd,
-                agent_name=agent_name,
-                bare=bare,
-                max_agents=max_agents,
-                dry_run=dry_run,
-            )
-        if cancel_scope.cancelled_caught:
-            raise LionTimeoutError(f"Flow timed out after {timeout}s")
-        return result
-    return await _run_flow_inner(
-        model_spec,
-        prompt,
-        with_synthesis=with_synthesis,
-        synthesis_model=synthesis_model,
-        max_concurrent=max_concurrent,
-        yolo=yolo,
-        verbose=verbose,
-        effort=effort,
-        theme=theme,
-        output_format=output_format,
-        save_dir=save_dir,
-        team_name=team_name,
-        cwd=cwd,
-        agent_name=agent_name,
-        bare=bare,
-        max_agents=max_agents,
-        dry_run=dry_run,
-    )
-
-
-async def _run_flow_inner(
-    model_spec: str,
-    prompt: str,
-    *,
-    with_synthesis: bool = False,
-    synthesis_model: str | None = None,
-    max_concurrent: int = 0,
-    yolo: bool = False,
-    verbose: bool = False,
-    effort: str | None = None,
-    theme: str | None = None,
-    output_format: str = "text",
-    save_dir: str | None = None,
-    team_name: str | None = None,
-    cwd: str | None = None,
-    agent_name: str | None = None,
-    bare: bool = False,
-    max_agents: int = 0,
-    dry_run: bool = False,
-) -> str:
-    """Inner flow logic (no timeout wrapper)."""
-    t0 = time.monotonic()
-
-    # Load orchestrator profile
-    orc_profile: AgentProfile | None = None
-    if agent_name:
-        orc_profile = load_agent_profile(agent_name)
-        if orc_profile.model and not model_spec:
-            model_spec = orc_profile.model
-        if orc_profile.effort and not effort:
-            effort = orc_profile.effort
-        if orc_profile.yolo and not yolo:
-            yolo = True
-
-    if not model_spec:
-        raise ValueError(
-            "Provide a model spec or use -a/--agent to load a profile with a model."
-        )
-
-    orc_imodel = build_imodel_from_spec(
-        model_spec,
-        yolo=yolo,
-        verbose=verbose,
-        effort_override=effort,
-        theme=theme,
-    )
-    if cwd:
-        orc_imodel.endpoint.config.kwargs.setdefault("repo", Path(cwd))
-
-    # ── Phase 0: Orchestrator plans the DAG ──────────────────────────
-    builder = OperationGraphBuilder("Flow")
-    orc_system = orc_profile.system_prompt if orc_profile else None
-    orc_branch = Branch(
-        chat_model=orc_imodel,
-        system=orc_system,
-        log_config=DataLoggerConfig(auto_save_on_exit=False),
-        name="orchestrator",
-    )
-    session = Session(default_branch=orc_branch)
-
-    # Build role roster for orchestrator guidance
-    from ._agents import list_agents
-
-    available_roles = list_agents()
-    if bare:
-        roles_guidance = (
-            f"Available roles: {', '.join(available_roles)}. "
-            f"All workers use model {model_spec}. "
-            f"Roles define behavioral focus only — model is fixed."
-        )
-    else:
-        role_details = []
-        for role in available_roles:
-            try:
-                rp = load_agent_profile(role)
-                rm = rp.model or model_spec
-                detail = f"{role} (model: {rm}"
-                if rp.effort:
-                    detail += f", effort: {rp.effort}"
-                detail += ")"
-                role_details.append(detail)
-            except FileNotFoundError:
-                role_details.append(f"{role} (model: {model_spec})")
-        roles_guidance = f"Available agents: {'; '.join(role_details)}."
-
-    budget_note = ""
-    if max_agents > 0:
-        budget_note = f"BUDGET: You may define at most {max_agents} agents total. "
-
-    plan_root = builder.add_operation(
-        "operate",
-        branch=orc_branch,
-        instruct=Instruct(
-            instruction=(
-                "Create a FlowPlan for the following task. "
-                "Define agents as a flat list with dependency edges via depends_on. "
-                "Each agent needs a short id (e.g. 'r1', 'a1'), a role, a concrete "
-                "instruction, and optional depends_on listing ids of agents whose "
-                "output it needs. Agents with no depends_on run immediately in parallel."
-            ),
-            context={"task": prompt},
-            guidance=(
-                f"{roles_guidance} "
-                f"{budget_note}"
-                "CRITICAL: You MUST produce your output ONLY via the structured "
-                "output fields (the FlowPlan). Do NOT use any provider-native "
-                "subagent or tool-spawning features (no Agent tool, no subprocess "
-                "spawning, no delegation tools). The ONLY correct way to define "
-                "workers is by filling in the FlowPlan structured output. "
-                "Use ONLY roles from the available list above — do not invent "
-                "custom role names. "
-                "Keep the agent count minimal — only add agents that provide "
-                "distinct value. Prefer fewer agents with clear instructions "
-                "over many agents with overlapping scope. "
-                "CONTROL NODES: Set control=true on an agent to make it a "
-                "flow control checkpoint. Control agents (typically critics) "
-                "review prior work and decide whether the flow should continue "
-                "with additional rounds. Place at most one control node, "
-                "and it should depend on the agents whose work it reviews. "
-                "Set synthesis=true if the task benefits from a final consolidated output."
-            ),
-        ),
-        field_models=[FLOW_PLAN_FIELDS],
-        reason=True,
-    )
-
-    if not verbose:
-        print("Planning DAG...", file=sys.stderr)
-
-    result0 = await session.flow(builder.get_graph())
-    t_plan = time.monotonic() - t0
-
-    plan_result = result0.get("operation_results", {}).get(plan_root)
-    plan: FlowPlan | None = getattr(plan_result, "plan", None)
-
-    if not plan or not plan.agents:
-        return "Orchestrator produced no flow plan."
-
-    if max_agents > 0 and len(plan.agents) > max_agents:
-        plan.agents = plan.agents[:max_agents]
-        if not verbose:
-            print(f"Plan truncated to {max_agents} agents (--max-agents)", file=sys.stderr)
-
-    if not verbose:
-        dag_lines = []
-        for a in plan.agents:
-            deps = f" ← {','.join(a.depends_on)}" if a.depends_on else ""
-            dag_lines.append(f"{a.id}:{a.role}{deps}")
-        print(
-            f"Plan done ({t_plan:.1f}s): {len(plan.agents)} agents: "
-            f"{' | '.join(dag_lines)}",
-            file=sys.stderr,
-        )
-
-    if plan.synthesis and not with_synthesis:
-        with_synthesis = True
-
-    # ── Dry run: dump plan and exit ─────────────────────────────────
-    if dry_run:
-        lines = [f"FlowPlan ({len(plan.agents)} agents, synthesis={plan.synthesis})"]
-        lines.append("")
-        for a in plan.agents:
-            ctrl = " [CONTROL]" if a.control else ""
-            deps = f"  depends_on: {', '.join(a.depends_on)}" if a.depends_on else ""
-            model_note = f"  model: {a.model}" if a.model else ""
-            lines.append(f"  {a.id}: {a.role}{ctrl}")
-            lines.append(f"    instruction: {a.instruction[:120]}...")
-            if deps:
-                lines.append(deps)
-            if model_note:
-                lines.append(model_note)
-            if a.guidance:
-                lines.append(f"    guidance: {a.guidance[:80]}...")
-            lines.append("")
-
-        # Show resolved models
-        lines.append("Model resolution:")
-        for a in plan.agents:
-            if bare:
-                rm = a.model or model_spec
-                lines.append(f"  {a.id}: {rm} (bare)")
-            else:
-                rm, rp = _resolve_worker_spec(a.role)
-                if a.model:
-                    rm = a.model
-                src = "plan" if a.model else ("profile" if rp else "default")
-                lines.append(f"  {a.id}: {rm} ({src})")
-        return "\n".join(lines)
-
-    # ── Build agent name registry ───────────────────────────────────
-    default_ms = parse_model_spec(model_spec)
-    all_agent_names: list[str] = []
-    name_counts: dict[str, int] = {}
-    for a in plan.agents:
-        base = a.role
-        name_counts[base] = name_counts.get(base, 0) + 1
-        if name_counts[base] > 1:
-            all_agent_names.append(f"{base}-{name_counts[base]}")
-        else:
-            all_agent_names.append(base)
-
-    team_data = None
-    if team_name:
-        team_data = _create_fanout_team(team_name, all_agent_names)
-        if not verbose:
-            print(f"Team '{team_name}' created ({team_data['id']})", file=sys.stderr)
-
-    # ── Helper: resolve model/system/branch for an agent spec ─────
-    def _make_worker_branch(a: FlowAgentSpec, idx: int) -> tuple[Branch, str]:
-        w_profile = None
-        if bare:
-            w_model = a.model or default_ms.model
-        else:
-            w_model, w_profile = _resolve_worker_spec(a.role)
-            if a.model:
-                w_model = a.model
-            elif not w_profile:
-                w_model = default_ms.model
-        w_effort = effort
-        if not bare and w_profile and w_profile.effort and not effort:
-            w_effort = w_profile.effort
-        w_yolo = yolo
-        if not bare and w_profile and w_profile.yolo:
-            w_yolo = True
-        w_imodel = build_imodel_from_spec(
-            w_model, yolo=w_yolo, verbose=verbose,
-            effort_override=w_effort, theme=theme,
-        )
-        if cwd:
-            w_imodel.endpoint.config.kwargs.setdefault("repo", Path(cwd))
-        wname = all_agent_names[idx]
-        w_system = None
-        if team_data:
-            teammates = [n for n in all_agent_names if n != wname]
-            roster_lines = ["- orchestrator (coordinator)"]
-            roster_lines += [f"- {t}" for t in teammates]
-            roster_lines.append(f"- **{wname}** (you)")
-            w_system = TEAM_WORKER_SYSTEM.format(
-                worker_name=wname, team_name=team_data["name"],
-                team_id=team_data["id"],
-                roster_text="\n".join(roster_lines),
-            )
-        elif not bare and w_profile and w_profile.system_prompt:
-            w_system = w_profile.system_prompt
-        else:
-            w_system = (
-                "You are a leaf worker agent. Complete your assigned task "
-                "directly. You may read files, use tools, and run commands "
-                "as needed. Do NOT spawn sub-agents or delegate further."
-            )
-        wb = Branch(
-            chat_model=w_imodel, system=w_system,
-            log_config=DataLoggerConfig(auto_save_on_exit=False), name=wname,
-        )
-        session.include_branches(wb)
-        return wb, w_model
-
-    # ── Build DAG: split regular agents and control nodes ──────────
-    spec_to_node: dict[str, str] = {}
-    agent_meta: dict[str, dict] = {}
-    agent_results: list[dict] = []
-    regular_agents = [a for a in plan.agents if not a.control]
-    control_agents = [a for a in plan.agents if a.control]
-
-    if not verbose:
-        ctrl_note = f" ({len(control_agents)} control)" if control_agents else ""
-        print(f"Executing DAG: {len(regular_agents)} agents{ctrl_note}...", file=sys.stderr)
-
-    # Build regular agent nodes
-    for idx, a in enumerate(plan.agents):
-        if a.control:
-            continue
-        w_branch, w_model = _make_worker_branch(a, idx)
-        dep_nodes = []
-        if a.depends_on:
-            for dep_id in a.depends_on:
-                if dep_id in spec_to_node:
-                    dep_nodes.append(spec_to_node[dep_id])
-        if not dep_nodes:
-            dep_nodes = [plan_root]
-        node_id = builder.add_operation(
-            "instruct", branch=w_branch, depends_on=dep_nodes,
-            instruction=a.instruction, guidance=a.guidance,
-            context=[{"original_task": prompt}],
-        )
-        spec_to_node[a.id] = node_id
-        agent_meta[node_id] = {
-            "name": all_agent_names[idx], "model": w_model,
-            "spec_id": a.id, "depends_on": a.depends_on or [],
-        }
-
-    # Execute regular agents
-    t_exec = time.monotonic()
-    conc = max_concurrent if max_concurrent > 0 else max(len(regular_agents), 1)
-    dag_result = await session.flow(
-        builder.get_graph(), max_concurrent=conc, verbose=verbose,
-    )
-    t_exec_elapsed = time.monotonic() - t_exec
-
-    op_results = dag_result.get("operation_results", {})
-    for a in regular_agents:
-        nid = spec_to_node[a.id]
-        meta = agent_meta[nid]
-        res = op_results.get(nid)
-        agent_results.append({
-            "id": a.id, "name": meta["name"], "model": meta["model"],
-            "depends_on": meta["depends_on"], "control": False,
-            "response": str(res) if res is not None else "(no response)",
-            "time_ms": t_exec_elapsed * 1000,
-        })
-
-    if not verbose:
-        print(f"DAG done ({t_exec_elapsed:.1f}s).", file=sys.stderr)
-
-    # ── Execute control nodes: instruct with prior results ─────────
-    max_rounds = 3
-    round_num = 0
-    for ca in control_agents:
-        idx = next(i for i, a in enumerate(plan.agents) if a.id == ca.id)
-        c_branch, c_model = _make_worker_branch(ca, idx)
-
-        artifacts = [f"[{r['id']} ({r['name']})]: {r['response']}" for r in agent_results]
-        dep_nodes = []
-        if ca.depends_on:
-            for dep_id in ca.depends_on:
-                if dep_id in spec_to_node:
-                    dep_nodes.append(spec_to_node[dep_id])
-        if not dep_nodes:
-            all_node_ids = list(spec_to_node.values())
-            dep_nodes = all_node_ids[-1:] if all_node_ids else [plan_root]
-
-        if not verbose:
-            print(f"Control [{ca.id}]: evaluating...", file=sys.stderr)
-
-        ctrl_node = builder.add_operation(
-            "instruct", branch=c_branch, depends_on=dep_nodes,
-            instruction=(
-                f"{ca.instruction}\n\n"
-                "Review all prior agent outputs and produce a verdict: "
-                "should_continue (bool), reason (str), and optional "
-                "next_steps (str) guidance if continuing."
-            ),
-            guidance=ca.guidance,
-            context=[{"original_task": prompt}, {"agent_outputs": artifacts}],
-            field_models=[FLOW_VERDICT_FIELDS],
-            reason=True,
-        )
-        spec_to_node[ca.id] = ctrl_node
-
-        t_ctrl = time.monotonic()
-        ctrl_result = await session.flow(builder.get_graph(), verbose=verbose)
-        t_ctrl_elapsed = time.monotonic() - t_ctrl
-
-        ctrl_res = ctrl_result.get("operation_results", {}).get(ctrl_node)
-        verdict: FlowControlVerdict | None = getattr(ctrl_res, "verdict", None)
-        verdict_text = str(ctrl_res) if ctrl_res is not None else "(no response)"
-
-        agent_results.append({
-            "id": ca.id, "name": all_agent_names[idx], "model": c_model,
-            "depends_on": ca.depends_on or [], "control": True,
-            "response": verdict_text,
-            "time_ms": t_ctrl_elapsed * 1000,
-        })
-
-        if not verbose:
-            cont = verdict.should_continue if verdict else False
-            print(
-                f"Control [{ca.id}] done ({t_ctrl_elapsed:.1f}s): "
-                f"continue={cont}",
-                file=sys.stderr,
-            )
-
-        # ── If verdict says continue: orchestrator re-plans ────────
-        if verdict and verdict.should_continue:
-            round_num += 1
-            if round_num >= max_rounds:
-                if not verbose:
-                    print(f"Max rounds ({max_rounds}) reached, stopping.", file=sys.stderr)
-                break
-
-            if not verbose:
-                print(f"Round {round_num + 1}: orchestrator re-planning...", file=sys.stderr)
-
-            replan_node = builder.add_operation(
-                "operate", branch=orc_branch, depends_on=[ctrl_node],
-                instruct=Instruct(
-                    instruction=(
-                        "The control node requested continuation. Plan additional agents "
-                        "to address the feedback. Return a FlowPlan with new agents only."
-                    ),
-                    context={
-                        "original_task": prompt,
-                        "prior_results": artifacts,
-                        "control_verdict": verdict_text,
-                        "next_steps_guidance": verdict.next_steps or "",
-                    },
-                    guidance=(
-                        f"{roles_guidance} "
-                        f"This is round {round_num + 1}. "
-                        "Add only the agents needed to address the control feedback. "
-                        "Do NOT re-run agents that already succeeded. "
-                        "Use ONLY the structured output. Do NOT spawn subagents."
-                    ),
-                ),
-                field_models=[FLOW_PLAN_FIELDS],
-                reason=True,
-            )
-
-            replan_result = await session.flow(builder.get_graph(), verbose=verbose)
-            replan_res = replan_result.get("operation_results", {}).get(replan_node)
-            next_plan: FlowPlan | None = getattr(replan_res, "plan", None)
-
-            if next_plan and next_plan.agents:
-                if not verbose:
-                    ids = ", ".join(a.id for a in next_plan.agents)
-                    print(f"Re-plan: {len(next_plan.agents)} new agents: {ids}", file=sys.stderr)
-
-                # Build and execute new agents
-                new_names = []
-                for na in next_plan.agents:
-                    base = na.role
-                    name_counts[base] = name_counts.get(base, 0) + 1
-                    new_names.append(f"{base}-{name_counts[base]}")
-                all_agent_names.extend(new_names)
-
-                for ni, na in enumerate(next_plan.agents):
-                    nb, nm = _make_worker_branch(na, len(all_agent_names) - len(new_names) + ni)
-                    nd = []
-                    if na.depends_on:
-                        for did in na.depends_on:
-                            if did in spec_to_node:
-                                nd.append(spec_to_node[did])
-                    if not nd:
-                        nd = [ctrl_node]
-                    nid = builder.add_operation(
-                        "instruct", branch=nb, depends_on=nd,
-                        instruction=na.instruction, guidance=na.guidance,
-                        context=[{"original_task": prompt}, {"prior_results": artifacts}],
-                    )
-                    spec_to_node[na.id] = nid
-                    agent_meta[nid] = {
-                        "name": new_names[ni], "model": nm,
-                        "spec_id": na.id, "depends_on": na.depends_on or [],
-                    }
-
-                t_new = time.monotonic()
-                new_result = await session.flow(
-                    builder.get_graph(), max_concurrent=conc, verbose=verbose,
-                )
-                t_new_elapsed = time.monotonic() - t_new
-                new_op = new_result.get("operation_results", {})
-                for na in next_plan.agents:
-                    nid = spec_to_node[na.id]
-                    meta = agent_meta[nid]
-                    res = new_op.get(nid)
-                    agent_results.append({
-                        "id": na.id, "name": meta["name"], "model": meta["model"],
-                        "depends_on": meta["depends_on"], "control": False,
-                        "response": str(res) if res is not None else "(no response)",
-                        "time_ms": t_new_elapsed * 1000,
-                    })
-                if not verbose:
-                    print(f"Round {round_num + 1} done ({t_new_elapsed:.1f}s).", file=sys.stderr)
-
-    # ── Synthesis ────────────────────────────────────────────────────
-    synthesis_result = None
-    if with_synthesis and agent_results:
-        synth_spec = synthesis_model or model_spec
-        synth_label = str(parse_model_spec(synth_spec))
-        if not verbose:
-            print(f"Synthesis [{synth_label}]...", file=sys.stderr)
-
-        all_node_ids = set(spec_to_node.values())
-        depended_on = set()
-        for a in plan.agents:
-            for d in (a.depends_on or []):
-                if d in spec_to_node:
-                    depended_on.add(spec_to_node[d])
-        leaf_nodes = list(all_node_ids - depended_on) or list(all_node_ids)
-
-        artifacts = [f"[{r['id']} ({r['name']})]: {r['response']}" for r in agent_results]
-        synth_node = builder.add_operation(
-            "instruct", branch=orc_branch, depends_on=leaf_nodes,
-            instruction=(
-                f"Synthesize all agent outputs into a final cohesive deliverable.\n\n"
-                f"Original task: {prompt}"
-            ),
-            context=artifacts,
-        )
-        t_synth = time.monotonic()
-        synth_result = await session.flow(builder.get_graph(), verbose=verbose)
-        t_synth_elapsed = time.monotonic() - t_synth
-        synth_res = synth_result.get("operation_results", {}).get(synth_node)
-        synthesis_result = {
-            "model": synth_label,
-            "response": str(synth_res) if synth_res is not None else "(no response)",
-            "time_ms": t_synth_elapsed * 1000,
-        }
-        if not verbose:
-            print(f"Synthesis done ({t_synth_elapsed:.1f}s).", file=sys.stderr)
-
-    # ── Output ───────────────────────────────────────────────────────
-    if output_format == "json":
-        output = _format_result_json(agent_results, synthesis_result)
-    else:
-        output = _format_flow_result_text(agent_results, synthesis_result)
-
-    # ── Save ─────────────────────────────────────────────────────────
-    if save_dir:
-        save_path = Path(save_dir)
-        save_path.mkdir(parents=True, exist_ok=True)
-        for r in agent_results:
-            p = save_path / f"{r['id']}_{r['name']}.md"
-            p.write_text(r["response"])
-        if synthesis_result:
-            (save_path / "synthesis.md").write_text(synthesis_result["response"])
-        if not verbose:
-            print(f"Saved to {save_path}", file=sys.stderr)
-
-    if team_data:
-        _post_results_to_team(team_data, agent_results, all_agent_names, synthesis_result)
-
-    # ── Persist branches ─────────────────────────────────────────────
-    orc_provider = orc_branch.chat_model.endpoint.config.provider
-    orc_branch_id = str(orc_branch.id)
-    orc_path = await acreate_path(
-        directory=LIONAGI_HOME / "logs" / "agents" / orc_provider,
-        filename=orc_branch_id,
-        file_exist_ok=True,
-    )
-    await orc_path.write_text(json_dumps(orc_branch.to_dict()))
-    save_last_branch_pointer(orc_provider, orc_branch_id)
-
-    worker_branch_ids = []
-    for branch in session.branches:
-        if str(branch.id) == orc_branch_id:
-            continue
-        w_provider = branch.chat_model.endpoint.config.provider
-        w_branch_id = str(branch.id)
-        w_path = await acreate_path(
-            directory=LIONAGI_HOME / "logs" / "agents" / w_provider,
-            filename=w_branch_id,
-            file_exist_ok=True,
-        )
-        await w_path.write_text(json_dumps(branch.to_dict()))
-        worker_branch_ids.append((w_provider, w_branch_id, branch.name))
-
-    t_total = time.monotonic() - t0
-    if not verbose:
-        print(f"\nTotal: {t_total:.1f}s", file=sys.stderr)
-
-    print(f'\n[orchestrator] li agent -r {orc_branch_id} "..."', file=sys.stderr)
-    for wp, wid, wn in worker_branch_ids:
-        print(f'[{wn}] li agent -r {wid} "..."', file=sys.stderr)
-
-    return output
-
-
-# ── CLI registration ──────────────────────────────────────────────────────
-
-
-def add_orchestrate_subparser(subparsers: argparse._SubParsersAction) -> None:
-    """Register `li orchestrate` (alias `li o`) with its sub-commands."""
-    orch = subparsers.add_parser(
-        "orchestrate",
-        aliases=["o"],
-        help="Multi-agent orchestration patterns.",
-        description="Orchestrate multiple agents in structured patterns.",
-    )
-    orch_sub = orch.add_subparsers(dest="orch_command", required=True)
-
-    fo = orch_sub.add_parser(
-        "fanout",
-        help="Fan-out N workers in parallel, optionally synthesize.",
-        description=(
-            "Orchestrator decomposes task into N agent requests, "
-            "fans out to workers, optionally synthesizes. "
-            "Effort can be embedded in model spec: claude/opus-4-7-high."
-        ),
-    )
-    fo.add_argument(
-        "model",
-        nargs="?",
-        default=None,
-        help=(
-            "Orchestrator model spec (provider/model-effort). "
-            "Also used as default worker model unless --workers specified. "
-            "Optional when -a/--agent provides a model."
-        ),
-    )
-    fo.add_argument("prompt", help="Task prompt for the orchestrator to decompose.")
-    fo.add_argument(
-        "-a",
-        "--agent",
-        metavar="NAME",
-        default=None,
-        help=(
-            "Load orchestrator profile from .lionagi/agents/<NAME>.md. "
-            "Profile provides system prompt, default model, effort, yolo. "
-            "CLI flags and positional model override profile settings."
-        ),
-    )
-
-    fo.add_argument(
-        "-n",
-        "--num-workers",
-        type=int,
-        default=3,
-        help="Number of workers (default: 3). Ignored if --workers set.",
-    )
-    fo.add_argument(
-        "--workers",
-        metavar="M1,M2,...",
-        default=None,
-        help="Comma-separated worker model specs (each can include effort).",
-    )
-    fo.add_argument(
-        "--max-concurrent",
-        type=int,
-        default=0,
-        help="Max concurrent workers (default: all).",
-    )
-    fo.add_argument(
-        "--with-synthesis",
-        nargs="?",
-        const=True,
-        default=False,
-        metavar="MODEL",
-        help="Enable synthesis. Bare flag uses orchestrator model; with arg uses that model.",
-    )
-    fo.add_argument(
-        "--synthesis-prompt",
-        default=None,
-        help="Custom synthesis instruction.",
-    )
-    fo.add_argument(
-        "--output",
-        choices=("text", "json"),
-        default="text",
-        help="Output format (default: text).",
-    )
-    fo.add_argument(
-        "--save",
-        metavar="DIR",
-        default=None,
-        help="Save outputs to directory.",
-    )
-
-    fo.add_argument(
-        "--team-mode",
-        nargs="?",
-        const="fanout",
-        default=None,
-        metavar="NAME",
-        help=(
-            "Create a persistent team for this fanout. Workers get team context "
-            "and results are posted as team messages. Bare flag uses 'fanout' as "
-            "team name; with arg uses that name."
-        ),
-    )
-
-    add_common_cli_args(fo)
-
-    # ── flow sub-command ─────────────────────────────────────────────
-    fl = orch_sub.add_parser(
-        "flow",
-        help="Auto-DAG pipeline: orchestrator plans DAG, engine executes.",
-        description=(
-            "Orchestrator analyzes the task, composes a DAG of agents "
-            "with dependency edges, and executes with automatic "
-            "parallelism where dependencies allow."
-        ),
-    )
-    fl.add_argument(
-        "model",
-        nargs="?",
-        default=None,
-        help="Orchestrator model spec. Optional when -a/--agent provides one.",
-    )
-    fl.add_argument("prompt", help="Task for the orchestrator to plan and execute.")
-    fl.add_argument(
-        "-a",
-        "--agent",
-        metavar="NAME",
-        default=None,
-        help="Load orchestrator profile from .lionagi/agents/<NAME>.md.",
-    )
-    fl.add_argument(
-        "--with-synthesis",
-        nargs="?",
-        const=True,
-        default=False,
-        metavar="MODEL",
-        help="Enable final synthesis. Bare flag uses orchestrator model.",
-    )
-    fl.add_argument(
-        "--max-concurrent",
-        type=int,
-        default=0,
-        help="Max concurrent agents within a phase (default: all).",
-    )
-    fl.add_argument(
-        "--output",
-        choices=("text", "json"),
-        default="text",
-        help="Output format (default: text).",
-    )
-    fl.add_argument(
-        "--save",
-        metavar="DIR",
-        default=None,
-        help="Save outputs to directory.",
-    )
-    fl.add_argument(
-        "--team-mode",
-        nargs="?",
-        const="flow",
-        default=None,
-        metavar="NAME",
-        help="Create a persistent team for this flow.",
-    )
-    fl.add_argument(
-        "--dry-run",
-        action="store_true",
-        help="Plan the DAG but don't execute. Shows agents, deps, and model resolution.",
-    )
-    fl.add_argument(
-        "--bare",
-        action="store_true",
-        help=(
-            "Ignore agent profiles — all workers use the CLI model spec. "
-            "Roles define behavioral focus only, no profile system prompts."
-        ),
-    )
-    fl.add_argument(
-        "--max-agents",
-        type=int,
-        default=0,
-        metavar="N",
-        help="Max total agents the orchestrator may plan (0 = unlimited).",
-    )
-    add_common_cli_args(fl)
-
-
-def run_orchestrate(args: argparse.Namespace) -> int:
-    """Dispatch orchestrate sub-commands."""
-    if args.orch_command == "fanout":
-        has_model = args.model is not None or args.agent is not None
-        if not has_model:
-            print(
-                "error: model or --agent is required",
-                file=sys.stderr,
-            )
-            return 1
-
-        synth = args.with_synthesis
-        with_synthesis = synth is not False
-        synthesis_model = synth if isinstance(synth, str) else None
-
-        try:
-            output = run_async(
-                _run_fanout(
-                    model_spec=args.model or "",
-                    prompt=args.prompt,
-                    num_workers=args.num_workers,
-                    workers_str=args.workers,
-                    with_synthesis=with_synthesis,
-                    synthesis_model=synthesis_model,
-                    synthesis_prompt=args.synthesis_prompt,
-                    max_concurrent=args.max_concurrent,
-                    yolo=args.yolo,
-                    verbose=args.verbose,
-                    effort=args.effort,
-                    theme=args.theme,
-                    output_format=args.output,
-                    save_dir=args.save,
-                    team_name=args.team_mode,
-                    cwd=args.cwd,
-                    timeout=args.timeout,
-                    agent_name=args.agent,
-                )
-            )
-        except LionTimeoutError as e:
-            print(str(e), file=sys.stderr)
-            return 1
-        if not args.verbose:
-            print(output)
-        return 0
-
-    if args.orch_command == "flow":
-        has_model = args.model is not None or args.agent is not None
-        if not has_model:
-            print("error: model or --agent is required", file=sys.stderr)
-            return 1
-
-        synth = args.with_synthesis
-        with_synthesis = synth is not False
-        synthesis_model = synth if isinstance(synth, str) else None
-
-        try:
-            output = run_async(
-                _run_flow(
-                    model_spec=args.model or "",
-                    prompt=args.prompt,
-                    with_synthesis=with_synthesis,
-                    synthesis_model=synthesis_model,
-                    max_concurrent=args.max_concurrent,
-                    yolo=args.yolo,
-                    verbose=args.verbose,
-                    effort=args.effort,
-                    theme=args.theme,
-                    output_format=args.output,
-                    save_dir=args.save,
-                    team_name=args.team_mode,
-                    cwd=args.cwd,
-                    timeout=args.timeout,
-                    agent_name=args.agent,
-                    bare=args.bare,
-                    max_agents=args.max_agents,
-                    dry_run=args.dry_run,
-                )
-            )
-        except LionTimeoutError as e:
-            print(str(e), file=sys.stderr)
-            return 1
-        if not args.verbose:
-            print(output)
-        return 0
-
-    print(f"Unknown orchestrate command: {args.orch_command}", file=sys.stderr)
-    return 1
