+++ lionagi/cli/orchestrate/_common.py
@@ -1,0 +1,163 @@
+# Copyright (c) 2023-2026, HaiyangLi <quantocean.li at gmail dot com>
+# SPDX-License-Identifier: Apache-2.0
+"""Shared models, formatting, team helpers, and worker resolution."""
+
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+from lionagi import Branch, FieldModel, json_dumps
+from lionagi.models import HashableModel
+from lionagi.operations.fields import Instruct
+from lionagi.protocols.generic.log import DataLoggerConfig
+
+from .._agents import AgentProfile, load_agent_profile
+from .._persistence import LIONAGI_HOME, save_last_branch_pointer
+from .._providers import build_imodel_from_spec, parse_model_spec
+from ..team import _now_iso, _save_team
+
+# ── Agent request model (structured output from orchestrator) ─────────────
+
+
+class AgentRequest(HashableModel):
+    """A single agent request generated by the orchestrator."""
+
+    instruct: Instruct
+    model: str | None = None
+
+
+AGENT_REQUEST_FIELDS = FieldModel(list[AgentRequest], name="agents")
+
+
+# ── Output formatting ─────────────────────────────────────────────────────
+
+
+def _format_result_text(
+    worker_results: list[dict],
+    synthesis_result: dict | None = None,
+) -> str:
+    lines = []
+    for i, w in enumerate(worker_results, 1):
+        n = len(worker_results)
+        lines.append(f"{'═' * 60}")
+        lines.append(f"  Worker {i}/{n}  [{w['model']}]")
+        lines.append(f"  {w['time_ms']:.0f}ms")
+        lines.append(f"{'═' * 60}")
+        lines.append(w.get("response", "(no response)"))
+        lines.append("")
+
+    if synthesis_result is not None:
+        lines.append(f"{'═' * 60}")
+        lines.append(f"  Synthesis  [{synthesis_result['model']}]")
+        lines.append(f"  {synthesis_result['time_ms']:.0f}ms")
+        lines.append(f"{'═' * 60}")
+        lines.append(synthesis_result.get("response", "(no response)"))
+        lines.append("")
+
+    return "\n".join(lines)
+
+
+def _format_result_json(
+    worker_results: list[dict],
+    synthesis_result: dict | None = None,
+) -> str:
+    out = {"workers": worker_results}
+    if synthesis_result is not None:
+        out["synthesis"] = synthesis_result
+    return json_dumps(out)
+
+
+# ── Team-mode guidance ────────────────────────────────────────────────────
+
+TEAM_WORKER_SYSTEM = """\
+You are **{worker_name}**, a specialist on team "{team_name}" (id: {team_id}).
+
+## Your team
+{roster_text}
+
+## How this works
+1. You receive a focused assignment from the orchestrator.
+2. Work independently — produce clear, actionable output.
+3. Your results are automatically shared with the team via `li team`.
+4. After this round, teammates or the orchestrator can follow up:
+   - `li team receive -t {team_id} --as {worker_name}` to read messages
+   - `li team send "..." -t {team_id} --to {worker_name}` to reply
+   - `li agent -r {{branch_id}} "follow-up"` to continue your session
+
+Focus on YOUR assignment. Be specific — your output is the deliverable.\
+"""
+
+
+def _create_fanout_team(
+    team_name: str,
+    worker_names: list[str],
+) -> dict:
+    from uuid import uuid4
+
+    team_id = uuid4().hex[:12]
+    members = ["orchestrator"] + worker_names
+    data = {
+        "id": team_id,
+        "name": team_name,
+        "members": members,
+        "messages": [],
+        "created_at": _now_iso(),
+    }
+    _save_team(data)
+    return data
+
+
+def _post_results_to_team(
+    team_data: dict,
+    worker_results: list[dict],
+    worker_names: list[str],
+    synthesis_result: dict | None = None,
+) -> None:
+    from uuid import uuid4
+
+    for wr, name in zip(worker_results, worker_names):
+        team_data["messages"].append(
+            {
+                "id": uuid4().hex[:12],
+                "from": name,
+                "to": ["*"],
+                "content": wr.get("response", "(no response)"),
+                "timestamp": _now_iso(),
+                "read_by": [],
+            }
+        )
+
+    if synthesis_result:
+        team_data["messages"].append(
+            {
+                "id": uuid4().hex[:12],
+                "from": "orchestrator",
+                "to": ["*"],
+                "content": f"[SYNTHESIS]\n{synthesis_result.get('response', '')}",
+                "timestamp": _now_iso(),
+                "read_by": [],
+            }
+        )
+
+    _save_team(team_data)
+
+
+# ── Worker resolution ────────────────────────────────────────────────────
+
+
+def _resolve_worker_spec(
+    token: str,
+) -> tuple[str, AgentProfile | None]:
+    """Resolve a worker token to (model_spec, profile_or_None).
+
+    If token contains '/' → model spec (e.g. 'codex/gpt-5.4').
+    Otherwise → try loading as agent profile name.
+    """
+    if "/" in token:
+        return token, None
+    try:
+        profile = load_agent_profile(token)
+        return profile.model or "codex/gpt-5.4", profile
+    except FileNotFoundError:
+        return token, None
